requirements.txt
Django==3.2.3
psycopg2-binary>=2.8
djangorestframework==3.12.4
djoser==2.1.0
gunicorn==20.1.0
django-crispy-forms
